requirements.txt
celery
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
python-resize-image==1.1.11
redis==2.10.5
requests==2.18.3
rq==0.8.1
urllib3==1.22
Werkzeug==0.12.2
